manager_agent/requirements.txt
google-cloud-aiplatform[adk,agent_engines]
google-cloud-firestore
google-auth
requests
python-dotenv
google-cloud-documentai